pyproject.toml
[project]
name = "hello-world-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-google-genai>=2.0.7",
    "python-dotenv>=1.0.1",
]
